Moved evaluation functions to new file evaluation.py
